+++ 2026-04/01_block_fetcher/store/ethdb/util.go
@@ -1,0 +1,7 @@
+package mdbxethdb
+
+import "github.com/erigontech/mdbx-go/mdbx"
+
+func isNotFound(err error) bool {
+	return mdbx.IsNotFound(err)
+}